+++ Changelog.md
@@ -4,6 +4,7 @@
 * use a single state atom in simulator
 * added perfect power factor to math
 * generalized quantum phase estimation and integrated in quantum-period-finding
+* fixed coprime selection in shor
 
 ## Version 0.1.0
 * implemented quantom states, quantum gates, gate registry and quantum cirquits